+++ inst/tinytest/test_unique_Postcodes.R
@@ -8,6 +8,9 @@ x <- c(x, NA_integer_)
 expect_equal(unique_Postcodes(x),
              sort(unique(x)))
 
+expect_equal(unique_Postcodes(c(1L, 800L), strict = TRUE), 800L)
+expect_equal(uniqueN_Postcodes(1:9999), healthyAddress:::get_N_POSTCODES())
+expect_equal(last(unique_Postcodes(1:9999)), healthyAddress:::get_MAX_POSTCODE())